+++ src/main/java/org/scubakay/dynamic_resource_pack/DynamicResourcePack.java
@@ -1,11 +1,11 @@
 package org.scubakay.dynamic_resource_pack;
 
-import de.maxhenkel.configbuilder.ConfigBuilder;
 import dev.kikugie.fletching_table.annotation.fabric.Entrypoint;
+import eu.midnightdust.lib.config.MidnightConfig;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 import org.scubakay.dynamic_resource_pack.command.ResourceCommand;
-import org.scubakay.dynamic_resource_pack.config.ModConfig;
+import org.scubakay.dynamic_resource_pack.config.Config;
 import org.scubakay.dynamic_resource_pack.util.ConfigFileHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -17,18 +17,11 @@ public class DynamicResourcePack implements ModInitializer {
     public static final String MOD_ID = "DynamicResourcePack";
     public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
 
-    public static ModConfig modConfig;
-
     @Override
     public void onInitialize() {
+        MidnightConfig.init(MOD_ID, Config.class);
         ConfigFileHandler.registerEvents();
         CommandRegistrationCallback.EVENT.register(ResourceCommand::register);
-
-        modConfig = ConfigBuilder.builder(ModConfig::new)
-                .path(getConfigFile())
-                .strict(true)
-                .saveAfterBuild(true)
-                .build();
     }
 
     public Path getConfigDirectory() {